Added external delivery

+++ hMailServer/Delivery/MessageDeliverer.cs
@@ -1,8 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using hMailServer.Core;
+using hMailServer.Core.Dns;
+using hMailServer.Core.Entities;
 using hMailServer.Core.Logging;
 using hMailServer.Entities;
 using hMailServer.Repository;
@@ -76,6 +79,7 @@ namespace hMailServer.Delivery
             var accountRepository = _container.GetInstance<IAccountRepository>();
             var messageRepository = _container.GetInstance<IMessageRepository>();
             var folderRepository = _container.GetInstance<IFolderRepository>();
+            var dnsClient = _container.GetInstance<IDnsClient>();
 
             message.NumberOfDeliveryAttempts++;
 
@@ -85,11 +89,27 @@ namespace hMailServer.Delivery
 
             try
             {
-                var localDelivery = new LocalDelivery(accountRepository, messageRepository, folderRepository);
+                var remainingRecipients = new List<Recipient>(message.Recipients);
 
-                deliveryResults.AddRange(await localDelivery.DeliverAsync(message));
+                var localDelivery = new LocalDelivery(accountRepository, messageRepository, folderRepository, _log);
+                deliveryResults.AddRange(await localDelivery.DeliverAsync(message, remainingRecipients.Where(recipient => recipient.AccountId != 0).ToList()));
 
-                await messageRepository.DeleteAsync(message);
+                var externalDelivery = new ExternalDelivery(messageRepository, dnsClient, _log);
+                deliveryResults.AddRange(await externalDelivery.DeliverAsync(message, remainingRecipients.Where(recipient => recipient.AccountId == 0).ToList()));
+
+                var failedRecipients =
+                    deliveryResults.Where(result => result.ReplyCodeSeverity == ReplyCodeSeverity.PermanentNegative ||
+                                                    (isLastAttempt && result.ReplyCodeSeverity == ReplyCodeSeverity.TransientNegative));
+
+                await SubmitBounceMessageAsync(message, failedRecipients);
+
+                var deliveryCompleted =
+                    deliveryResults.Any(result => result.ReplyCodeSeverity == ReplyCodeSeverity.TransientNegative);
+
+                if (isLastAttempt  || !deliveryCompleted)
+                {
+                    await messageRepository.DeleteAsync(message);
+                }
             }
             catch (Exception ex)
             {
@@ -117,6 +137,37 @@ namespace hMailServer.Delivery
                 }
 
             }
+        }
+
+        private async Task SubmitBounceMessageAsync(Message message, IEnumerable<DeliveryResult> failedRecipients)
+        {
+            if (string.IsNullOrWhiteSpace(message.From))
+                return;
+
+            if (IsMailerDaemonAddress(message.From))
+                return;
+
+            // TODO: Dont' hardcode this.
+            string bounceMessage = string.Format(@"Your message did not reach some or all of the intended recipients.
+
+   Sent: %MACRO_SENT%
+   Subject: %MACRO_SUBJECT%
+
+The following recipient(s)could not be reached:
+
+%MACRO_RECIPIENTS%
+
+hMailServer");
+
+            throw new NotImplementedException();
+
+        }
+
+        private bool IsMailerDaemonAddress(string address)
+        {
+            var mailbox = EmailAddressParser.GetMailbox(address);
+
+            return mailbox.Equals("MAILER-DAEMON", StringComparison.InvariantCultureIgnoreCase);
         }
     }
 }
